:zap: refactor: refactor desktop and mobile to improve mobile performance
