+++ src/app/market/(desktop)/layout/index.tsx
@@ -5,7 +5,7 @@ import { PropsWithChildren } from 'react';
 import ResponsiveLayout from '@/components/ResponsiveLayout';
 import { useIsMobile } from '@/hooks/useIsMobile';
 
-import Mobile from '../layout.mobile';
+import Mobile from '../../(mobile)/layout.mobile';
 import Desktop from './Desktop';
 
 const Market = ({ children }: PropsWithChildren) => (
