+++ src/app/market/components/AgentDetail/AgentInfo/Loading.tsx
