+++ src/app/market/components/AgentDetail/AgentInfo/TokenTag.tsx
