+++ src/app/market/components/AgentDetail/AgentInfo/style.ts
