+++ src/app/market/components/AgentDetail/Mobile.tsx
