+++ src/app/market/components/AgentDetail/index.tsx
@@ -1,0 +1,10 @@
+import { memo } from 'react';
+import { Flexbox } from 'react-layout-kit';
+
+import AgentInfo from './AgentInfo';
+
+export const AgentDetailContent = memo(() => (
+  <Flexbox>
+    <AgentInfo />
+  </Flexbox>
+));
