+++ src/app/market/layout.tsx
@@ -3,8 +3,8 @@ import { PropsWithChildren } from 'react';
 import ResponsiveLayout from '@/components/ResponsiveLayout';
 import { isMobileDevice } from '@/utils/responsive';
 
-import Desktop from './layout.desktop';
-import Mobile from './layout.mobile';
+import Desktop from './(desktop)/layout';
+import Mobile from './(mobile)/layout.mobile';
 
 export default ({ children }: PropsWithChildren) => (
   <ResponsiveLayout Desktop={Desktop} Mobile={Mobile} isMobile={isMobileDevice}>
